missing episodes on kodi
This script will look at all series that kodi has stored on mysql, look up the episode information on tvdb, and then check for missing episodes in your library.

Report will come back with possible duplicates and those that are missing.
Any episode on TVDB that does NOT have an air date will also appear in the list.
Episodes that are combined will also show as missing eg - video file is called S1,E23-24. TVDB will have it as two separates so will show as missing.

This will need you to provide sqldetails for the mysql database that kodi stores its information into.

NOTE: MySQLdb python must be installed for connectivity to work.